refactor: tweak builtin-plugin
