+++ packages/rolldown/src/builtin-plugin/constructor.ts
@@ -1,0 +1,9 @@
+import type { BindingBuiltinPluginName } from '../binding';
+
+export class BuiltinPlugin {
+  constructor(
+    public name: BindingBuiltinPluginName,
+    // NOTE: has `_` to avoid conflict with `options` hook
+    public _options?: unknown,
+  ) {}
+}
